VICTORY!!!  got the JOGL version working as well as the Java3D version!

+++ Torquescape-core/src/org/andork/torquescape/control/UVIntersector.java
@@ -4,35 +4,35 @@ package org.andork.torquescape.control;
 
 public class UVIntersector
 {
-	final float[ ]	t			= new float[ 3 ];
+	final double[ ]	t			= new double[ 3 ];
 	final int[ ]	edgeIndices	= new int[ 3 ];
-	float			u;
-	float			v;
+	double			u;
+	double			v;
 	
-	int compare( float t0 , float t1 )
+	int compare( double t0 , double t1 )
 	{
-		if( Float.isNaN( t0 ) || Float.isInfinite( t0 ) )
+		if( Double.isNaN( t0 ) || Double.isInfinite( t0 ) )
 		{
-			return Float.isNaN( t1 ) || Float.isInfinite( t1 ) ? 0 : 1;
+			return Double.isNaN( t1 ) || Double.isInfinite( t1 ) ? 0 : 1;
 		}
 		else
 		{
 			if( t0 < 0 )
 			{
-				return t1 < 0 ? Float.compare( t1 , t0 ) : 1;
+				return t1 < 0 ? Double.compare( t1 , t0 ) : 1;
 			}
 			else
 			{
-				return t1 < 0 ? -1 : Float.compare( t0 , t1 );
+				return t1 < 0 ? -1 : Double.compare( t0 , t1 );
 			}
 		}
 	}
 	
-	void swapIfNecessary( float[ ] t , int[ ] sides , int i , int j )
+	void swapIfNecessary( double[ ] t , int[ ] sides , int i , int j )
 	{
 		if( i > j != compare( t[ i ] , t[ j ] ) > 0 )
 		{
-			float swap = t[ i ];
+			double swap = t[ i ];
 			t[ i ] = t[ j ];
 			t[ j ] = swap;
 			
@@ -42,11 +42,11 @@ public class UVIntersector
 		}
 	}
 	
-	public void intersect( float u0 , float v0 , float u , float v )
+	public void intersect( double u0 , double v0 , double u , double v )
 	{
-		t[ 0 ] = u >= 0 ? Float.NaN : -u0 / u;
-		t[ 1 ] = v >= 0 ? Float.NaN : -v0 / v;
-		t[ 2 ] = u + v <= 0 ? Float.NaN : ( 1 - u0 - v0 ) / ( u + v );
+		t[ 0 ] = u >= 0 ? Double.NaN : -u0 / u;
+		t[ 1 ] = v >= 0 ? Double.NaN : -v0 / v;
+		t[ 2 ] = u + v <= 0 ? Double.NaN : ( 1 - u0 - v0 ) / ( u + v );
 		
 		edgeIndices[ 0 ] = 2;
 		edgeIndices[ 1 ] = 0;
